pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "v1nizin", email = "viniwick@outlook.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

